backend-generation/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
google-genai==1.40.0
pillow==10.4.0
moviepy==2.1.2
imageio-ffmpeg==0.6.0
